lucid-voice/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
faster-whisper>=1.0.0
piper-tts>=1.2.0
python-multipart>=0.0.9
numpy>=1.26.0
